+++ packages/gatsby-theme-aio/gatsby-node.js
@@ -133,6 +133,18 @@ exports.createResolvers = ({ createResolvers, addFrontmatterType }) => {
         type: '[String]',
         resolve: addFrontmatterType
       },
+      contributor_name: {
+        type: 'String',
+        resolve: addFrontmatterType
+      },
+      contributor_link: {
+        type: 'String',
+        resolve: addFrontmatterType
+      },
+      edition: {
+        type: 'String',
+        resolve: addFrontmatterType
+      },
       openAPISpec: {
         type: 'String',
         resolve: addFrontmatterType