+++ packages/gatsby-theme-aio/src/components/MDXFilter/index.js
@@ -14,6 +14,8 @@ import React, { Children, cloneElement, useContext } from 'react';
 import { withPrefix } from 'gatsby';
 import { MDXProvider } from '@mdx-js/react';
 import { css } from '@emotion/react';
+import '@spectrum-css/divider';
+import '@spectrum-css/actiongroup';
 import Context from '../Context';
 import {
   layoutColumns,
@@ -36,6 +38,8 @@ import { Contributors } from '../Contributors';
 import { Feedback } from '../Feedback';
 import { GitHubActions } from '../GitHubActions';
 import { Breadcrumbs } from '../Breadcrumbs';
+import { Attribution } from '../Attribution';
+import { Edition } from '../Edition';
 import { OnThisPage } from '../OnThisPage';
 import { NextSteps } from '../NextSteps';
 import { NextPrev } from '../NextPrev';
@@ -148,6 +152,13 @@ export default ({ children, pageContext, query }) => {
     const selectedPageSiblings = findSelectedPageSiblings(location?.pathname, siteMetadata?.subPages);
     const { nextPage, previousPage } = findSelectedPageNextPrev(location?.pathname, siteMetadata?.subPages);
 
+    // Attribution
+    const contributorName = pageContext?.frontmatter?.contributor_name;
+    const contributorLink = pageContext?.frontmatter?.contributor_link;
+
+    // Edition
+    const edition = pageContext?.frontmatter?.edition;
+
     // OnThisPage
     const componentPathObj = allSitePage?.nodes.find(({ path }) => withPrefix(path) === location?.pathname);
     const componentPath = componentPathObj?.component ?? '';
@@ -163,8 +174,11 @@ export default ({ children, pageContext, query }) => {
     const pagePath = componentPath.replace(/.*\/src\/pages\//g, '');
 
     // Breadcrumbs
-    const hideBreadcrumbNav = pageContext?.frontmatter?.hideBreadcrumbNav !== undefined ? pageContext.frontmatter.hideBreadcrumbNav : false;
-    if(typeof hideBreadcrumbNav != "boolean"){ throw new Error("hideBreadcrumbNav is not a boolean. Correct use hideBreadcrumbNav: true"); }
+    const hideBreadcrumbNav =
+      pageContext?.frontmatter?.hideBreadcrumbNav !== undefined ? pageContext.frontmatter.hideBreadcrumbNav : false;
+    if (typeof hideBreadcrumbNav != 'boolean') {
+      throw new Error('hideBreadcrumbNav is not a boolean. Correct use hideBreadcrumbNav: true');
+    }
     const { home } = siteMetadata;
     const pathWithRootFix = rootFix(location.pathname);
     const pagesWithRootFix = rootFixPages(siteMetadata?.pages);
@@ -320,6 +334,8 @@ export default ({ children, pageContext, query }) => {
                     <div
                       css={css`
                         margin-left: auto;
+                        display: flex;
+                        align-items: center;
 
                         @media screen and (max-width: ${DESKTOP_SCREEN_WIDTH}) {
                           margin-left: 0;
@@ -330,6 +346,13 @@ export default ({ children, pageContext, query }) => {
                     </div>
                   </div>
                 )}
+                <div
+                  css={css`
+                    display: block;
+                  `}>
+                  {edition && <Edition name={edition} />}
+                  {contributorLink && <Attribution name={contributorName} link={contributorLink} />}
+                </div>
 
                 {filteredChildren}
 
@@ -344,8 +367,7 @@ export default ({ children, pageContext, query }) => {
                       flex-wrap: wrap;
                       align-items: center;
                       justify-content: space-between;
-                      margin-top: var(--spectrum-global-dimension-size-800);
-                      margin-bottom: var(--spectrum-global-dimension-size-400);
+                      margin-bottom: var(--spectrum-global-dimension-size-200);
                     `}>
                     <div>
                       <Contributors
@@ -373,9 +395,7 @@ export default ({ children, pageContext, query }) => {
                   </div>
                 )}
               </div>
-
               {hasOnThisPage && <OnThisPage tableOfContents={tableOfContents} />}
-
               {resourcesChild && resourcesChild}
             </div>
           </div>